requirements.txt
blinker==1.7.0
Bootstrap-Flask==2.2.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email-validator==2.1.0
Faker==22.0.0
Flask==2.3.3
Flask-Avatars==0.2.2
Flask-Dropzone==1.6.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.0.5
flask-whooshee==0.8.2
Flask-WTF==1.1.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==25.0
pathtools==0.1.2
pillow==10.2.0
pillow-avif-plugin==1.4.1
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9.1
six==1.16.0
SQLAlchemy==2.0.25
text-unidecode==1.3
typing_extensions==4.15.0
urllib3==2.5.0
watchdog==3.0.0
Werkzeug==2.3.7
Whoosh==2.7.4
WTForms==3.0.1
